pyproject.toml
[project]
name = "aiblog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "pydantic-settings>=2.11.0",
    "google-generativeai>=0.7.2",
    "google-genai>=1.50.1",
    "alembic>=1.17.2",
    "uvicorn>=0.38.0",
    "psycopg2-binary>=2.9.11",
]
